Complete destroy assignment

+++ app/controllers/comments_controller.rb
@@ -12,7 +12,7 @@ class CommentsController < ApplicationController
 
     if @comment.save
       flash[:notice] = "Comment was saved."
-      redirect_to [@post]
+      redirect_to [@post.topic, @post]
     else
       flash[:error] = "There was an error saving the comment. Please try again."
       render :new
@@ -20,15 +20,17 @@ class CommentsController < ApplicationController
   end
 
   def destroy
-    @comment = @post.comments.find(params[:id])
+  @comment = @post.comments.find(params[:id])
 
+    authorize @post
     authorize @comment
+
     if @comment.destroy
       flash[:notice] = "Comment was removed."
-      redirect_to [@post]
+      redirect_to [@post.topic, @post]
     else
       flash[:error] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post]
+      redirect_to [@post.topic, @post]
     end
   end
